+++ service/cziti/windns/windns.go
@@ -29,10 +29,7 @@ import (
 var log = pfxlog.Logger()
 
 func ResetDNS() {
-	log.Infof("skipping ResetDNS for now")
-
-	log.Infof("resetting dns...")
-	log.Info("restoring dns to original-ish state")
+	log.Info("resetting dns to original-ish state")
 
 	script := `Get-NetIPInterface | ForEach-Object { Set-DnsClientServerAddress -InterfaceIndex $_.ifIndex -ResetServerAddresses }`
 
@@ -47,8 +44,6 @@ func ResetDNS() {
 }
 
 func GetUpstreamDNS() []string {
-	ResetDNS()
-
 	script := `Get-DnsClientServerAddress | ForEach-Object { $_.ServerAddresses } | Sort-Object | Get-Unique`
 
 	cmd := exec.Command("powershell", "-Command", script)
@@ -77,46 +72,6 @@ func GetUpstreamDNS() []string {
 }
 
 func ReplaceDNS(ips []net.IP) {
-
-	/*
-	ifaces, err := net.Interfaces()
-	if err != nil {
-		fmt.Print(fmt.Errorf("localAddresses: %+v\n", err.Error()))
-		return
-	}
-	for _, i := range ifaces {
-		addrs, err := i.Addrs()
-		if err != nil {
-			fmt.Print(fmt.Errorf("localAddresses: %+v\n", err.Error()))
-			continue
-		}
-		for _, a := range addrs {
-			switch v := a.(type) {
-			case *net.IPAddr:
-				fmt.Printf("%v : %s (%s)\n", i.Name, v, v.IP.DefaultMask())
-
-			case *net.IPNet:
-				fmt.Printf("%v : %s [%v/%v]\n", i.Name, v, v.IP, v.Mask)
-			}
-
-		}
-	}
-
-	var names []string
-	for _, i := range ips {
-		names = append(names, i.String())
-	}
-	addresses := strings.Join(names, ",")
-
-	*/
-	/*
-	script := fmt.Sprintf(
-		`Get-NetIPInterface | ForEach-Object { Set-DnsClientServerAddress -InterfaceIndex $_.ifIndex -ServerAddresses %s }`,
-		addresses)
-	*/
-
-
-
 	script := `$dnsinfo=Get-DnsClientServerAddress
 
 # see https://docs.microsoft.com/en-us/dotnet/api/system.net.sockets.addressfamily
@@ -132,7 +87,7 @@ foreach ($dns in $dnsinfo)
         $dnsServers=$dns.ServerAddresses
         $ArrList=[System.Collections.ArrayList]@($dnsServers)
         if(($dnsServers -ne $null) -and ($dnsServers.Contains("::1")) ) {
-            echo ($dns.InterfaceAlias + " IPv6 already contains ::1")
+            # uncomment when debugging echo ($dns.InterfaceAlias + " IPv6 already contains ::1")
         } else {
             $ArrList.Insert(0,"::1")
         }
@@ -142,7 +97,7 @@ foreach ($dns in $dnsinfo)
         $dnsServers=$dns.ServerAddresses
         $ArrList=[System.Collections.ArrayList]@($dnsServers)
         if(($dnsServers -ne $null) -and ($dnsServers.Contains("127.0.0.1")) ) {
-            echo ($dns.InterfaceAlias + " IPv4 already contains 127.0.0.1")
+            # uncomment when debugging echo ($dns.InterfaceAlias + " IPv4 already contains 127.0.0.1")
         } else {
             $ArrList.Insert(0,"127.0.0.1")
         }