+++ release-notes.md
@@ -1,3 +1,18 @@
+# Release 0.0.17
+
+* What's New
+
+* Bug fixes
+  * [#59](https://github.com/openziti/desktop-edge-win/issues/59) - too many services blocked service from accepting connections from the UI
+  * [#61](https://github.com/openziti/desktop-edge-win/issues/61) - identity shutdown needs to be on the uv loop (issue with forgetting identities)
+  * [#63](https://github.com/openziti/desktop-edge-win/issues/63) - when service restarts and UI reconnects clear identities and let the service repopulate the UI
+  * [#65](https://github.com/openziti/desktop-edge-win/issues/65) - set the MTU based on the value reported from the interface
+
+
+# Release 0.0.16
+
+- tracking lost
+
 # Release 0.0.15
 
 * What's New